fix frames; fix *_var stuff; fix JsshSocket.
fix frames on Firefox class and FFElement.
get rid of old commented code.
change a bunch of (document|window|whatever)_var stuff to use document_object, etc.
get rid of window_object and use browser_window_object or content_window_object as appropriate instead.
replace a bunch more javascript with ruby.
JsshSocket checks if it's just getting a prompt before switching to short timeout (which is now zero) so stuff doesn't unintentionally get left on the socket. work stuff to avoid blank strings that make it wait the full timeout for data that's not going to come. in the future I may work on switching jssh prompt types so that this can be avoided.