change my mind on element-locating so that relocating is possible; add Firefox#updated_at.
change some of how elements are specified and located back to old way of creating an Element and locating it rather than locating and creating. this way if the document changes, the Element can re-locate itself. add updated_at method of Watir::Firefox that tracks when this re-locating might need to happen.
work on code related to element locating.
remove a bunch of element_collection stuff as now that just happens with Arrays. a problem with this is that things suddenly change to 0-based indexing rather than 1.
fix frame so that its dom_object refers to its document (where getElementsByTagName and getElementById are defined, so it works as a container).
add some methods to select_list, fix aliasing to old methods.