element.rb: stop checking if if container_collection_method is defined on container_module before creating the method - nothing else defines container collection methods anymore. create a new variant of container collection methods: child_{collection method}, for example child_divs will return divs that are immediate children of the container; child_elements will return elements that are immediate children of the container. uses childNodes of the dom object for this.
fix a bug in element_collection where element_object_candidates was looking at container's element_object rather than containing_object. add complaint in ElementCollection#last if something is already using extra[:candidates] since #last overrides that. may fix this in the future to use extra[:index]=:last for this instead.

specifier.rb: add check for nodeType==1 before considering an object to be a candidate for an element (ELEMENT_NODE nodeType is 1).
this maybe should be in Element class's add_specifier. but that doesn't inherit, so it'd have to be on all the add_specifiers, and not really any point. (maybe it should inherit? that'd be tricky.)