rewrite IEElement#element_object_exists?
when an element stops existing in IE, in some cases the WIN32OLE raises WIN32OLERuntimeError when you try to use it. this was handled previously.

in some cases, the WIN32OLE continues to work fine, but its parentNode is nil. this was handled previously.

in some cases (when a whole chunk of the DOM is replaced by javascript), it seems that the WIN32OLE continues to work fine, its parentNode appears to point to a valid DOM node - but neither the WIN32OLE nor its parent are in fact a part of the DOM tree underneath the current document. this was not handled previously. so, we have to iterate up through parentNodes to the document and check if we encounter the correct document. if we don't, the element_object WIN32OLE refers to an element that does not exist on the page. so element_object_exists? returns false.