fix criteria for using getElementById and getElementsByName in specifier_candidates.
getElementById was using @index which didn't mean anything since this moved out to a module; fixed that. altered specifier_candidates to take an optional third argument to indicate whether we want the first one, and thus getElementById may be used. changed call to specifier_candidates to use third argument.

getElementsByTagName was being used to look for any element, which didn't work in IE because name is an expando attribute on any element for which it's not a valid attribute. restricted specifier_candidates to only use getElementsByTagName to look when specifiers specify tagNames that correspond to elements on which a name attribute is valid.