rework what files get required by what, where.
vapir-common.rb, vapir-ie.rb, and vapir-firefox.rb do the work of requiring what is needed for the libraries. vapir/ie.rb, vapir/ff.rb, vapir.rb, etc require these.
add vapir/common.rb, which just calls to vapir-common.rb

rename common_*.rb files (elements, container, modal_dialog) - so named because they would be named the same as ie files, and the directory in the path was called "watir" in both cases, so ambiguous. now they are called "vapir-common" and "vapir-ie", so not ambiguous, so files can be named the same.

change the main require files (vapir-firefox.rb and vapir-ie.rb) to not require absolutely everything - defer requiring stuff to stuff that actually needs it. the classes just require the browser-specific browser class, elements, and version; requiring the rest of the files that are used is deferred to files that actually need them.

likewise change individual files to add requires for all of the other stuff that they do require; previously expected to already be required.

add 'elements.rb' to both browsers to require all of the elements that are defined.

delete some empty files, no longer used; drop any requires of them.

rename page-container to page_container for consistency.