initial implementation of #visible_text
iterates down through nodes of the DOM and collects the text nodes which are visible. the work is done by #visible_text_nodes.
also adding #text_nodes, which was a first trial on iterating down through the dom. in firefox, element.text_nodes.join('') is equal to el.text. not always in IE - differences happen on some whitespace or nbsp chars.