add support for how/what specifiers on element collections. for example, browser.divs(:class => /someclass/) returns an ElementCollection which represents divs with the css class matching /someclass/. browser.elements(:xpath, "//input[@type='checkbox']") - well, you get the idea.
internally to ElementCollection, change #candidates to #element_objects, since the return value of that method is element objects that are actually to be used - not just candidates that may be used.
update that function to respect @how, behaving similiarly to Element#locate to get candidates and match them against a given how/what.
change things in ElementCollection that instantiate Elements to pass how/what along if that's defined.
change ElementCollection#find / #detect to fall back to Enumerable's #find/#detect method if @how is defined, since the behavior of using the how=:custom specifier is incompatible with passing along a specified 'how'.