rework screenshot methods to add support for firefox.
move WinWindow screenshot code out to common so that both IE and Firefox can use it. change second argument of IE's #screen_capture method to be an options hash.
add a cross-platform implementation of screenshots for Firefox which relies on capturing the data from a canvas element. this code is adapted from Selenium.