add default to #handling_existence_failure to check existence before proceeding.
since generally when calling #handling_existence_failure, you are unsure about existence, you'll want to immediately check that a thing still exists and handle the failure if it doesn't. so, making this the default.
changing #exists? not to use the default behavior as that would result in an infinite loop.
changing IE's overloaded #handling_existence_failure to rely more on the base method (now aliased for that usage)