initial implementation of config framework.
this will implement a high degree of granularity of control over basically every option which should be configurable in vapir.
initial implementation has stubs for base defaults, env variables to override those, and Vapir.config to override env in-process.

Vapir::Configuration is a class which basically wraps around a hash of key/value pairs but does so heirarchically so that more specific config can override broader ones (such as env overriding defaults and Vapir.config overriding env).

Vapir::Configuration::Option represents a configuration option and stores information on the key and what is valid for the value, in the form of @validator which is used to check values when they are set. this may be used for other information about configuration options in the future.